Refactor breadcrumbs
