--- app/cms/route-presets/cadence/sidebar/index.yml
@@ -1,37 +1,0 @@
-schema: "../sidebar-schema.json"
-# ---------------------------------
-# /cadence
-sidebar:            
-  - Cadence:
-# -------------------------------
-    - label: Language Reference,
-    - href: language
-
-    - label: Tutorial
-    - href: tutorial
-
-  - Developer Guides:
-# -------------------------------
-    - label: Introduction to Cadence
-    - href: index
-
-    - label: Cadence Design Patterns
-    - href: design-patterns
-
-    - label: Contract Upgrades with Incompatible Changes
-    - href: contract-upgrades
-
-    - label: Cadence Anti-Patterns
-    - href: anti-patterns
-
-    - label: msg․sender Considered Harmful
-    - href: msg-sender
-
-    - label: Measuring Time in Cadence
-    - href: measuring-time
-
-    - label: Migration Guide
-    - href: migration-guide
-
-    - label: JSON-Cadence Data Interchange Format
-    - href: json-cadence-spec
