--- app/cms/route-presets/cadence/sidebar/language.yml
@@ -1,12 +1,0 @@
-# /cadence/language
-sidebar: 
-- Cadence Language:
-# -------------------------------
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
