--- app/cms/route-presets/cadence/sidebar/tutorial.yml
@@ -1,12 +1,0 @@
-# /cadence/tutorial
-sidebar: 
-- Cadence Tutorial:
-# -------------------------------
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
-  - lable: ok
-    href: ok
