+++ app/ui/design-system/src/lib/Pages/InternalPage/useInternalBreadcrumbs.ts
@@ -1,11 +1,4 @@
-import { capitalCase } from "change-case"
 import { useMemo } from "react"
-import { displayNames } from "~/cms/schema"
-import {
-  FIRST_ROUTE_MAP,
-  isSecondRoute,
-  SecondRoute,
-} from "~/constants/repos/contents-structure"
 import { InternalSidebarSectionItem } from "../../Components/InternalSidebar"
 
 export type UseInternalBreadcrumbsOptions = {
@@ -41,32 +34,5 @@ export const useInternalBreadcrumbs = ({
   rootUrl = "/",
 }: UseInternalBreadcrumbsOptions) =>
   useMemo(() => {
-    const breadcrumbs = [{ href: rootUrl, name: "Home" }]
-
-    var basePath = `${rootUrl}${contentPath}`
-
-    if (isSecondRoute(contentPath)) {
-      const firstRouteName = FIRST_ROUTE_MAP[contentPath as SecondRoute]!
-
-      breadcrumbs.push({
-        name: displayNames[firstRouteName] || capitalCase(firstRouteName),
-        href: `${rootUrl}${firstRouteName}`,
-      })
-      breadcrumbs.push({
-        name: contentDisplayName,
-        href: `${rootUrl}${firstRouteName}/${contentPath}`,
-      })
-      basePath = `${rootUrl}${firstRouteName}/${contentPath}`
-    } else {
-      breadcrumbs.push({ name: contentDisplayName, href: basePath })
-    }
-
-    if (activeItem) {
-      breadcrumbs.push({
-        name: activeItem.label,
-        href: `${basePath}/${activeItem.href}`,
-      })
-    }
-
-    return breadcrumbs
-  }, [activeItem, contentDisplayName, contentPath, rootUrl])
+    return []
+  }, [])
